Re-organise homepage structure
