+++ frontend/src/components/homepage-page-handler/schedule-preview-section.tsx
@@ -1,0 +1,3 @@
+export const SchedulePreviewSection = () => {
+  return <div>abc</div>;
+};
