--- frontend/src/components/sponsors-section/types.ts
@@ -1,5 +1,0 @@
-export type Sponsor = {
-  image?: string | null;
-  link?: string | null;
-  name: string;
-};
